chore: impl integration caching
